Updating ImageMagick to 7.1.1-19

+++ Magick++/lib/Image.cpp
@@ -2,7 +2,7 @@
 //
 // Copyright Bob Friesenhahn, 1999, 2000, 2001, 2002, 2003
 //
-// Copyright @ 2013 ImageMagick Studio LLC, a non-profit organization
+// Copyright @ 1999 ImageMagick Studio LLC, a non-profit organization
 // dedicated to making software imaging solutions freely available.
 //
 // Implementation of Image
@@ -764,7 +764,7 @@ void Magick::Image::fileName(const std::string &fileName_)
   modifyImage();
 
   max_length=sizeof(image()->filename)-1;
-  fileName_.copy(image()->filename,max_length);
+  fileName_.copy(image()->filename,(size_t) max_length);
   if ((ssize_t) fileName_.length() > max_length)
     image()->filename[max_length]=0;
   else
@@ -816,7 +816,7 @@ void Magick::Image::fillPattern(const Image &fillPattern_)
 
 Magick::Image Magick::Image::fillPattern(void) const
 {
-  // FIXME: This is inordinately innefficient
+  // FIXME: This is inordinately inefficient
   const MagickCore::Image
     *tmpTexture;
 
@@ -1482,7 +1482,7 @@ void Magick::Image::strokePattern(const Image &strokePattern_)
 
 Magick::Image Magick::Image::strokePattern(void) const
 {
-  // FIXME: This is inordinately innefficient
+  // FIXME: This is inordinately inefficient
   const MagickCore::Image 
     *tmpTexture;
 
@@ -4210,14 +4210,13 @@ void Magick::Image::roll(const Geometry &roll_)
   ThrowImageException;
 }
 
-void Magick::Image::roll(const size_t columns_,const size_t rows_)
+void Magick::Image::roll(const ssize_t columns_,const ssize_t rows_)
 {
   MagickCore::Image
     *newImage;
 
   GetPPException;
-  newImage=RollImage(constImage(),static_cast<ssize_t>(columns_),
-    static_cast<ssize_t>(rows_),exceptionInfo);
+  newImage=RollImage(constImage(),columns_, rows_,exceptionInfo);
   replaceImage(newImage);
   ThrowImageException;
 }
